--- mod/mods/CrowdControl/ccver
@@ -1,1 +1,0 @@
-1.1.0
