--- mod/mods/CrowdControl/manifest.json
@@ -1,15 +1,0 @@
-{
-  "id": "CrowdControl",
-  "name": "Crowd Control",
-  "version": "1.1.0",
-  "description": [
-      "Crowd Control allows your livestream viewers", 
-      "on any platform to interact with your game.",
-      "Twitch, YouTube, TikTok and more!",
-"http://crowdcontrol.live"
-  ],
-  "author": "The Crowd Control Team",
-  "load_before": [],
-"load_after": [],
-"commands": []
-}
